fix notes. create index.tsx for pages
