+++ packages/client/src/components/button/button.scss
@@ -66,6 +66,7 @@
 
 .button--select {
   justify-content: flex-end;
+  padding-right: 16px;
 
   &:after {
     content: url("data:image/svg+xml,%3Csvg width='16' height='9' viewBox='0 0 16 9' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M15 1L8 8L1 1' stroke='%233B4F7D' strokeWidth='1.8' strokeLinecap='round' strokeLinejoin='round'/%3E%3C/svg%3E");
