+++ packages/client/src/components/checkbox/checkbox.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import './checkbox.scss';
+
+interface ICheckBox {
+  classes?: string ;
+  text?: string;
+}
+
+export function CheckBox(props: ICheckBox) {
+  return (
+    <div className={'checkbox ' + props.classes}>
+      <input type="checkbox" className="checkbox__input" id="last-word" name="last-word" />
+      <label className="checkbox__label" htmlFor="last-word">{props.text}</label>
+    </div>
+  )
+}
+
+
