+++ packages/client/src/components/form-field/form-field.scss
@@ -32,13 +32,3 @@
 .form__fields--center {
   align-items: center;
 }
-
-.form__fild--row {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-}
-
-.form__fild--half {
-  width: 50%;
-}
