+++ packages/client/src/components/index.ts
@@ -4,3 +4,4 @@ export * from './form-field/form-field';
 export * from './intro/intro';
 export * from './page-navigation/page-navigation';
 export * from './breadcrumbs/breadcrumbs';
+export * from './checkbox/checkbox';
