+++ packages/client/src/pages/index.tsx
@@ -1,0 +1,9 @@
+export * from './main';
+export * from './login';
+export * from './sign-up';
+export * from './new-game';
+export * from './profile';
+export * from './rules';
+export * from './leaders';
+export * from './forum';
+export * from './round';
