+++ packages/client/src/scss/base/_scaffolding.scss
@@ -72,3 +72,7 @@ path {
 .invisible {
   visibility: hidden;
 }
+
+.col-6 {
+  width: 50%;
+}
