+++ packages/client/src/scss/base/_variables.scss
@@ -43,6 +43,6 @@ $button-shadow-hover: inset -4px -4px 4px rgba(0, 0, 0, 0.1);
 $menu-item-shadow: 4px 4px 40px 0px rgba(178, 193, 240, 0.7);
 $menu-item-pressed-shadow: 10px 10px 20px 0px rgba(178, 193, 240, 0.4) inset;
 $checkbox-shadow: 2px 2px 4px rgba(114, 142, 171, 0.1), -6px -6px 20px white, 4px 4px 20px rgba(111, 140, 176, 0.41);
-$breadcrumbs-shodow: inset -2px -2px 4px white, inset 2px 2px 4px #bac3df;
+$breadcrumbs-shadow: inset -2px -2px 4px white, inset 2px 2px 4px #bac3df;
 $breadcrumbs-shadow-hover:inset 2px 2px 4px#bac3df, inset 2px 2px 4px white;// Screen width
 $mobile: 768px;
